:sparkles: Namespaces: Project Input
