+++ src/app.ts
@@ -6,66 +6,9 @@
 /// <reference path="components/base-component.ts" />
 /// <reference path="components/project-item.ts" />
 /// <reference path="components/project-list.ts" />
+/// <reference path="components/project-input.ts" />
 
 namespace App {
-
-
-  // ProjectInput Class
-  class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
-    titleInputElement: HTMLInputElement;
-    descriptionInputElement: HTMLInputElement;
-    peopleInputElement: HTMLInputElement;
-
-    constructor() {
-      super("project-input", "app", true, "user-input");
-      this.titleInputElement = this.element.querySelector("#title")! as HTMLInputElement;
-      this.descriptionInputElement = this.element.querySelector("#description")! as HTMLInputElement;
-      this.peopleInputElement = this.element.querySelector("#people")! as HTMLInputElement;
-      this.configure();
-    }
-
-    renderContent() { }
-
-    configure() {
-      this.element.addEventListener("submit", this.submitHandler);
-    }
-
-    private gatherUserInput(): [string, string, number] | void {
-      const enteredTitle = this.titleInputElement.value;
-      const enteredDescription = this.descriptionInputElement.value;
-      const enteredPeople = this.peopleInputElement.value;
-
-      const titleValidatable: Validatable = {value: enteredTitle, required: true};
-      const descriptionValidatable: Validatable = {value: enteredDescription, required: true, minLength: 5};
-      const peopleValidatable: Validatable = {value: +enteredPeople, required: true, min: 1};
-
-      if (!validate(titleValidatable) || !validate(descriptionValidatable) || !validate(peopleValidatable)) {
-        alert("Invalid input, please try again!");
-        return;
-      } else {
-        return [enteredTitle, enteredDescription, +enteredPeople];
-      }
-    }
-
-    private clearInputs() {
-      this.titleInputElement.value = "";
-      this.descriptionInputElement.value = "";
-      this.peopleInputElement.value = "";
-    }
-
-    @autobind
-    private submitHandler(event: Event) {
-      event.preventDefault();
-      const userInput = this.gatherUserInput();
-      if (Array.isArray(userInput)) {
-        const [title, description, people] = userInput;
-        projectState.addProject(title, description, people);
-        this.clearInputs();
-      }
-    }
-
-  }
-
   new ProjectInput();
   new ProjectList('active');
   new ProjectList('finished');
